requirements.txt
streamlit
PyPDF2
textstat
pyspellchecker
